+++ aiohttp_client_cache/response.py
@@ -1,12 +1,17 @@
+# TODO: CachedResponse may be better as a non-slotted subclass of ClientResponse.
+#     Will look into this when working on issue #67.
+import asyncio
 import json
 from datetime import datetime
 from http.cookies import SimpleCookie
 from logging import getLogger
 from typing import Any, Dict, Iterable, List, Mapping, Optional, Tuple, Union
+from unittest.mock import Mock
 
 import attr
 from aiohttp import ClientResponse, ClientResponseError, hdrs, multipart
 from aiohttp.client_reqrep import ContentDisposition, MappingProxyType, RequestInfo
+from aiohttp.streams import StreamReader
 from aiohttp.typedefs import RawHeaders, StrOrURL
 from multidict import CIMultiDict, CIMultiDictProxy, MultiDict, MultiDictProxy
 from yarl import URL
@@ -56,7 +61,7 @@ class CachedResponse:
     _links: LinkItems = attr.ib(factory=list)
     cookies: SimpleCookie = attr.ib(default=None)
     created_at: datetime = attr.ib(factory=datetime.utcnow)
-    encoding: str = attr.ib(default=None)
+    encoding: str = attr.ib(default='utf-8')
     expires: Optional[datetime] = attr.ib(default=None)
     raw_headers: RawHeaders = attr.ib(factory=tuple)
     real_url: StrOrURL = attr.ib(default=None)
@@ -66,18 +71,19 @@ class CachedResponse:
     @classmethod
     async def from_client_response(cls, client_response: ClientResponse, expires: datetime = None):
         """Convert a ClientResponse into a CachedReponse"""
-        # Response may not have been read yet, if fetched by something other than CachedSession
-        if not client_response._released:
-            await client_response.read()
-
         # Copy most attributes over as is
         copy_attrs = set(attr.fields_dict(cls).keys()) - EXCLUDE_ATTRS
         response = cls(**{k: getattr(client_response, k) for k in copy_attrs})
 
-        # Set some remaining attributes individually
+        # Read response content, and reset StreamReader on original response
+        if not client_response._released:
+            await client_response.read()
         response._body = client_response._body
-        response._links = [(k, _to_str_tuples(v)) for k, v in client_response.links.items()]
+        client_response.content = CachedStreamReader(client_response._body)
+
+        # Set remaining attributes individually
         response.expires = expires
+        response.links = client_response.links
         response.real_url = client_response.request_info.real_url
 
         # The encoding may be unset even if the response has been read, and
@@ -85,7 +91,7 @@ class CachedResponse:
         try:
             response.encoding = client_response.get_encoding()
         except (RuntimeError, TypeError):
-            response.encoding = 'utf-8'
+            pass
 
         if client_response.history:
             response.history = (
@@ -94,8 +100,8 @@ class CachedResponse:
         return response
 
     @property
-    def _released(self):
-        return True
+    def content(self) -> StreamReader:
+        return CachedStreamReader(self._body)
 
     @property
     def content_disposition(self) -> Optional[ContentDisposition]:
@@ -117,6 +123,7 @@ class CachedResponse:
         """Get headers as an immutable, case-insensitive multidict from raw headers"""
 
         def decode_header(header):
+            """Decode an individual (key, value) pair"""
             return (
                 header[0].decode('utf-8', 'surrogateescape'),
                 header[1].decode('utf-8', 'surrogateescape'),
@@ -143,14 +150,14 @@ class CachedResponse:
         items = [(k, _to_url_multidict(v)) for k, v in self._links]
         return MultiDictProxy(MultiDict([(k, MultiDictProxy(v)) for k, v in items]))
 
+    @links.setter
+    def links(self, value: Mapping):
+        self._links = [(k, _to_str_tuples(v)) for k, v in value.items()]
+
     @property
     def ok(self) -> bool:
         """Returns ``True`` if ``status`` is less than ``400``, ``False`` if not"""
-        try:
-            self.raise_for_status()
-            return True
-        except ClientResponseError:
-            return False
+        return self.status < 400
 
     @property
     def request_info(self) -> RequestInfo:
@@ -174,8 +181,8 @@ class CachedResponse:
     def raise_for_status(self):
         if self.status >= 400:
             raise ClientResponseError(
-                self.request_info,  # type: ignore  # These types are interchangeable
-                tuple(),
+                self.request_info,
+                self.history,
                 status=self.status,
                 message=self.reason,
                 headers=self.headers,
@@ -185,12 +192,53 @@ class CachedResponse:
         """Read response payload."""
         return self._body
 
-    def release(self):
-        """No-op function for compatibility with ClientResponse"""
-
     async def text(self, encoding: Optional[str] = None, errors: str = "strict") -> str:
         """Read response payload and decode"""
         return self._body.decode(encoding or self.encoding, errors=errors)
+
+    # No-op/placeholder properties and methods that don't apply to a CachedResponse, but provide
+    # compatibility with aiohttp.ClientResponse
+    # ----------
+
+    @property
+    def _released(self):
+        return True
+
+    @property
+    def connection(self):
+        return None
+
+    async def __aenter__(self) -> 'CachedResponse':
+        return self
+
+    async def close(self):
+        pass
+
+    async def wait_for_close(self):
+        pass
+
+    def release(self):
+        pass
+
+    async def start(self):
+        pass
+
+    async def terminate(self):
+        pass
+
+
+class CachedStreamReader(StreamReader):
+    """A StreamReader loaded from previously consumed response content. This feeds cached data into
+    the stream so it can support all the same behavior as the original stream: async iteration,
+    chunked reads, etc.
+    """
+
+    def __init__(self, body: bytes = None):
+        body = body or b''
+        protocol = Mock(_reading_paused=False)
+        super().__init__(protocol, limit=len(body), loop=asyncio.get_event_loop())
+        self.feed_data(body)
+        self.feed_eof()
 
 
 AnyResponse = Union[ClientResponse, CachedResponse]