+++ HISTORY.md
@@ -1,12 +1,16 @@
 # History
 
 ## 0.5.0 (2021-TBD)
+[See all issues & PRs for v0.4](https://github.com/JWCook/aiohttp-client-cache/milestone/4?closed=1)
+
 * Add a filesystem backend
+* Add support for streaming requests
 * Add `RedisBackend.close()` method 
 * Add `MongoDBPickleCache.values()` method that deserializes items
 * Allow `BaseCache.has_url()` and `delete_url()` to take all the same parameters as `create_key()`
 * Improve normalization for variations of URLs & request parameters
 * Fix handling of request body when it has already been serialized
+* Add some missing no-op methods to `CachedResponse` for compatibility with `ClientResponse`
 
 ---
 ### 0.4.3 (2021-07-27)